+++ utils/forest_utils/src/lib.rs
@@ -1,5 +1,6 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
+
 pub extern crate const_format;
 
 pub mod db;
@@ -7,4 +8,5 @@ pub mod encoding;
 pub mod io;
 pub mod json;
 pub mod macros;
+pub mod monitoring;
 pub mod net;